requirements.txt
pyusb==1.2.1
Pillow==8.4.0